requirements.txt
aiohappyeyeballs   
aiohttp            
aiosignal          
annotated-doc      
annotated-types    
anyio              
asttokens          
async-timeout      
attrs             
certifi            
cffi               
charset-normalizer 
comm               
contourpy          
cryptography       
cycler             
debugpy            
decorator          
Deprecated         
distro             
exceptiongroup     
executing          
fastapi            
fonttools          
frozenlist         
genai              
google-auth        
google-genai       
h11                
httpcore           
httpx              
idna              
img2pdf            
ipykernel          
ipython            
jedi               
jupyter_client     
jupyter_core       
kiwisolver         
lxml               
matplotlib         
matplotlib-inline  
multidict          
narwhals           
nest-asyncio       
numpy              
ollama             
openai            
opencv-python      
packaging         
pandas             
parso              
pdf2image          
pexpect            
pikepdf            
pillow             
pip                
platformdirs       
plotly             
prompt_toolkit     
propcache          
psutil             
ptyprocess         
pure_eval          
pyasn1             
pyasn1_modules     
pycparser        
pydantic           
pydantic_core      
Pygments           
PyMuPDF            
pyparsing          
pypdf              
pyperclip          
pytesseract        
python-dateutil    
python-docx        
python-dotenv      
pytz               
pyzmq              
regex              
reportlab          
requests           
rsa                
setuptools         
six                
sniffio            
stack-data         
starlette          
tabulate           
tenacity           
tiktoken           
tornado            
tqdm              
traitlets          
typing_extensions  
typing-inspection  
tzdata             
urllib3            
wcwidth            
websockets         
wrapt              
yarl       
        